DONE!
Completed testing for after midnight and ensuring valid user input. Assume that the earliest morning start is 6am. This means the latest a restaurant could say they're open is 5:59am.

Because if you're looking for a restaurant that closed after 6am, that doesn't really make sense.

This was fun working on!